job_agent_demo/requirements.txt
streamlit>=1.43,<2
requests>=2.32,<3
python-dotenv>=1.0,<2

